Fix frame pipelining issues, implement N frame buffering

+++ src/DX12Engine/Rendering/Queues/CommandQueue.h
@@ -22,17 +22,17 @@ namespace DX12Engine
 		void InsertWaitForQueueFence(CommandQueue* otherQueue, UINT fenceValue);
 		void InsertWaitForQueue(CommandQueue* otherQueue);
 
-		void WaitForFenceCPUBlocking(UINT fenceValue);
+		void WaitForFenceCPUBlocking(UINT64 fenceValue);
 		void WaitForIdle() { WaitForFenceCPUBlocking(m_NextFenceValue - 1); }
 
 		Microsoft::WRL::ComPtr<ID3D12CommandQueue> GetCommandQueue() { return m_CommandQueue; }
 
-		UINT PollCurrentFenceValue();
-		UINT GetLastCompletedFence() { return m_LastCompletedFenceValue; }
-		UINT GetNextFenceValue() { return m_NextFenceValue; }
+		UINT64 PollCurrentFenceValue();
+		UINT64 GetLastCompletedFence() { return m_LastCompletedFenceValue; }
+		UINT64 GetNextFenceValue() { return m_NextFenceValue; }
 		Microsoft::WRL::ComPtr<ID3D12Fence> GetFence() { return m_Fence; }
 
-		UINT ExecuteCommandList();
+		UINT64 ExecuteCommandList();
 		ID3D12GraphicsCommandList* GetCommandList() { return m_CommandList.Get(); }
 		void ResetCommandAllocatorAndList();
 		void ResetCommandList();
